refactor: domain 패키지 game 및 board로 분리
