+++ src/main/java/domain/board/BoardFactory.java
@@ -1,6 +1,5 @@
-package domain;
+package domain.board;
 
-import domain.strategy.*;
 import domain.vo.Position;
 
 import java.util.Comparator;
