+++ src/main/java/domain/board/Palace.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.board;
 
 import domain.vo.Position;
