+++ src/main/java/domain/board/Piece.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.board;
 
 import domain.strategy.MoveStrategy;
 import domain.vo.Position;
