+++ src/main/java/domain/board/Team.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.board;
 
 public enum Team {
     CHU("초"),
