+++ src/main/java/domain/game/GameType.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 public enum GameType {
     NEW,
