+++ src/main/java/domain/game/Status.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 public enum Status {
     PLAYING,
