+++ src/main/java/domain/game/Turn.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.game;
+
+import domain.board.Team;
 
 public class Turn {
