+++ src/test/java/domain/TurnTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.board.Team;
+import domain.game.Turn;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
