+++ src/test/java/domain/strategy/CannonMoveStrategyTest.java
@@ -1,6 +1,10 @@
 package domain.strategy;
 
 import domain.*;
+import domain.board.Board;
+import domain.board.Piece;
+import domain.board.Team;
+import domain.board.Type;
 import domain.vo.Position;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
